Updated paths after code was moved to ckeditor5-core.
